+++ ssh-web/src/utils/router.tsx
@@ -12,9 +12,9 @@ import { Information } from '../pages/Information';
 import { Manage } from '../pages/Information/Manage';
 import { PromiseTicket } from '../pages/Promise';
 import { Request } from '../pages/Information/Request';
-import { Account } from '../pages/Account';
 import { IntroductionAccountPage } from '../pages/IntroductionAccount';
 import { AccountItemPage } from '../pages/AccountItem';
+import { Account } from '../pages/Account';
 
 export const PathNames: IPathNames = {
   HOME: {
@@ -92,7 +92,7 @@ export const router = createBrowserRouter([
       },
       {
         path: '/account',
-        element: <Account />,
+        element: <Account isParent={true} />,
       },
       {
         path: '/account/introduction',
